Add files via upload
Has been created three classes, Cliente, Carro and Aluguel.
Those classes are linked by an .CPP and a .H file, where the .H file is included in the main file.
The main file contains some functions thats used in the program. Next step in the atualization project of this little problem is expand those functions to a extra class.
I dont create an data base beacause its not necessary in this exercise. But can be implemented. Maybe in the future.
The data os the Clientes, Carros and Alugueis are saved in a .TXT file, with theys respectives names.